+++ app/gourmet-place-api/src/main/kotlin/com/gourmet/service/place/api/PlaceApplication.kt
@@ -1,0 +1,11 @@
+package com.gourmet.service.place.api
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class PlaceApplication
+
+fun main(args: Array<String>) {
+    runApplication<PlaceApplication>(*args)
+}
