+++ app/gourmet-place-api/src/main/kotlin/com/gourmet/service/place/api/payload/GetAllPlacesResponse.kt
@@ -1,0 +1,20 @@
+package com.gourmet.service.place.api.payload
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies
+import com.fasterxml.jackson.databind.annotation.JsonNaming
+import com.gourmet.service.place.core.usecase.dto.GetAllPlacesData
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
+data class GetAllPlacesResponse private constructor(
+    val id: String,
+    val name: String,
+    val thumbnail: String?
+) {
+    companion object {
+        fun fromData(data: GetAllPlacesData) = GetAllPlacesResponse(
+            id = data.id,
+            name = data.name,
+            thumbnail = data.thumbnail
+        )
+    }
+}
