+++ app/gourmet-place-api/src/test/kotlin/com/gourmet/service/place/api/PlaceApplicationTests.kt
@@ -1,0 +1,12 @@
+package com.gourmet.service.place.api
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class PlaceApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+}
