+++ common/gourmet-common-type/src/main/kotlin/com/gourmet/service/common/type/Geospatial.kt
