+++ common/gourmet-common-type/src/test/kotlin/com/gourmet/service/common/type/GeospatialPointTest.kt
