+++ core/gourmet-place-core/src/main/kotlin/com/gourmet/service/place/core/usecase/PlaceRepository.kt
@@ -1,0 +1,8 @@
+package com.gourmet.service.place.core.usecase
+
+import com.gourmet.service.place.core.domain.Place
+import reactor.core.publisher.Flux
+
+interface PlaceRepository {
+    fun getAllPlaces(): Flux<Place>
+}
