+++ core/gourmet-place-core/src/test/kotlin/com/gourmet/service/place/core/usecase/PlaceServiceTest.kt
@@ -1,0 +1,4 @@
+package com.gourmet.service.place.core.usecase
+
+class PlaceServiceTest {
+}
