+++ infra/gourmet-place-persistence/src/test/kotlin/com/gourmet/service/place/infra/persistence/PlaceRepositoryTest.kt
@@ -1,0 +1,3 @@
+package com.gourmet.service.place.infra.persistence
+
+class PlaceRepositoryTest
