--- module-api-place/src/main/kotlin/com/gourmet/service/api/place/PlaceApplication.kt
@@ -1,11 +1,0 @@
-package com.gourmet.service.api.place
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class PlaceApplication
-
-fun main(args: Array<String>) {
-    runApplication<PlaceApplication>(*args)
-}
