--- module-api-place/src/test/kotlin/com/gourmet/service/api/place/PlaceApplicationTests.kt
@@ -1,12 +1,0 @@
-package com.gourmet.service.api.place
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class PlaceApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-}
