--- module-core-place/src/main/kotlin/com/gourmet/service/core/place/domain/Place.kt
@@ -1,16 +1,0 @@
-package com.gourmet.service.core.place.domain
-
-import com.gourmet.service.common.type.GeospatialPoint
-import java.time.OffsetDateTime
-
-data class Place(
-    val id: String?,
-    val name: String,
-    val addr: String,
-    val thumbnail: String?,
-    val images: List<String>,
-    val location: GeospatialPoint,
-    val tags: List<String>,
-    val createdTime: OffsetDateTime?,
-    val updatedTime: OffsetDateTime?
-)
