--- module-core-place/src/main/kotlin/com/gourmet/service/core/place/usecase/PlaceRepository.kt
@@ -1,8 +1,0 @@
-package com.gourmet.service.core.place.usecase
-
-import com.gourmet.service.core.place.domain.Place
-import reactor.core.publisher.Flux
-
-interface PlaceRepository {
-    fun getAllPlaces(): Flux<Place>
-}
