--- module-core-place/src/test/kotlin/com/gourmet/service/core/place/usecase/PlaceServiceTest.kt
@@ -1,4 +1,0 @@
-package com.gourmet.service.core.place.usecase
-
-class PlaceServiceTest {
-}
