--- module-infra-place/src/test/kotlin/com/gourmet/service/infra/place/persistence/MockedPlaceRepositoryTest.kt
@@ -1,3 +1,0 @@
-package com.gourmet.service.infra.place.persistence
-
-class MockedPlaceRepositoryTest
